Start a new container for every combination of board+shield.
- Clean up build directory for every combo
- Copy CONFIG_* settings as artifacts to output directory
- Set and create OUTPUT_DIR within HOST_CONFIG_DIR

+++ scripts/build_board_matrix.sh
@@ -12,9 +12,8 @@ fi
 compile_board () {
     board="$1"
     shield="$2"
-    # [ -n "$shield" ] && extra_args="-DSHIELD=\"$shield\"" || extra_args=
-    # echo "$artifact_name $BUILD_DIR $extra_args"
     LOGFILE="$LOG_DIR/zmk_build_$artifact_name.log"
+
     echo -en "\n${GREEN}Building $1... ${NC}"
     west build -s "$DOCKER_ZMK_DIR/app" -d "build/$BUILD_DIR" -b "$1" "$WEST_OPTS" \
         -- -DZMK_CONFIG="$CONFIG_DIR" "$extra_args" -Wno-dev > "$LOGFILE" 2>&1
@@ -22,7 +21,7 @@ compile_board () {
     if [[ $? -eq 0 ]]
     then
         echo
-        echo "Build log saved to \"$LOGFILE\"."
+        echo "Build log saved to \"$LOGFILE\"."; echo
         echo "💪 ${GREEN}$artifact_name was built succesfully!${NC}"
         if [[ -f $DOCKER_ZMK_DIR/app/build/$BUILD_DIR/zephyr/zmk.uf2 ]]
         then
@@ -30,17 +29,23 @@ compile_board () {
         else
             TYPE="bin"
         fi
-        echo
+
         # OUTPUT="$OUTPUT_DIR/$1-zmk.$TYPE"
-        OUTPUT="$DOCKER_CONFIG_DIR/$artifact_name.$TYPE"
+        OUTPUT="$DOCKER_CONFIG_DIR/$OUTPUT_DIR/$artifact_name.$TYPE"
         # TODO: Use git tags to create a better extension than .bak
-        echo "💾 Renaming & backing up firmware file..."
+        echo "💾 Renaming & backing up firmware file & CONFIG_* file..."
         [[ -f $OUTPUT ]] && [[ ! -L $OUTPUT ]] && mv "$OUTPUT" "$OUTPUT.bak"
-        cp "$DOCKER_ZMK_DIR/app/build/$BUILD_DIR/zephyr/zmk.$TYPE" "$DOCKER_ZMK_DIR/app/build/$BUILD_DIR/zephyr/$artifact_name.$TYPE"
+
+        # Also copy CONFIG_* settings for possible debugging purposes.
+        CONFIG_OUTPUT="$DOCKER_CONFIG_DIR/$OUTPUT_DIR/CONFIG_$artifact_name"
+        grep -v -e "^#" -e "^$" "$DOCKER_ZMK_DIR/app/build/$BUILD_DIR/zephyr/.config" | \
+            sort > "$CONFIG_OUTPUT"
+
+        cp "$DOCKER_ZMK_DIR/app/build/$BUILD_DIR/zephyr/zmk.$TYPE" \
+            "$DOCKER_ZMK_DIR/app/build/$BUILD_DIR/zephyr/$artifact_name.$TYPE"
+
         cp "$DOCKER_ZMK_DIR/app/build/$BUILD_DIR/zephyr/zmk.$TYPE" "$OUTPUT" \
             && echo "⚙️ Copied firmware file to host folder."
-             printf "_______________________________________\n"
-
     else
         echo
         cat "$LOGFILE"
@@ -59,12 +64,12 @@ cd .. || exit
 
 OLD_WEST="/root/west.yml.old"
 if [[ ! -f "${DOCKER_ZMK_DIR}"/.west/config ]]; then
-    printf "🚀 Initiating the app... 🚀\n"
+    printf "🚀 Initializing the app... 🚀\n\n"
     west init -l app/
     cd "${DOCKER_ZMK_DIR}/app" || exit
     west update
 else
-    printf "✅ app is already initialized!\n"
+    printf "✅ app is already initialized!\n\n"
 fi
 
 cd "$DOCKER_ZMK_DIR/app" || exit
@@ -73,33 +78,23 @@ if [[ -f $OLD_WEST ]]; then
     md5_old=$(md5sum $OLD_WEST | cut -d' ' -f1)
 fi
 if [[ $md5_old != $(md5sum ./west.yml | cut -d' ' -f1) ]]; then
-    printf "\n🚀 app/west.yml has changed, Running 'west update' 🚀\n"
-    echo
+    printf "\n🚀 Found fresh app/west.yml, Running 'west update' 🚀\n\n"
     cp ./west.yml $OLD_WEST
     west update
 else
-    echo "✅ ${DOCKER_ZMK_DIR}/app/west.yml hasn't changed!"
+    printf "✅ ${DOCKER_ZMK_DIR}/app/west.yml hasn't changed!\n\n"
 fi
 
 west zephyr-export
 
 
-readarray -t board_shields < <(yaml2json "$DOCKER_CONFIG_DIR"/build.yaml | jq -c -r '.include[]')
-
-for line in "${board_shields[@]}"; do
-    read -ra arr <<< "${line//,/ }"
-    board=$(echo "${arr[0]}" | cut -d ':' -f 2 | sed 's/["{}}]//g')
-    shield=$(echo "${arr[1]}" | cut -d ':' -f 2 | sed 's/["{}}]//g')
-
-    artifact_name=${shield:+$shield-}${board}-zmk
-    extra_args=${shield:+-DSHIELD="$shield"}
-    BUILD_DIR="${artifact_name}_$SUFFIX"
+artifact_name=${shield:+$shield-}${board}-zmk
+extra_args=${shield:+-DSHIELD="$shield"}
+BUILD_DIR="${artifact_name}_$SUFFIX"
+if [ -d "$DOCKER_ZMK_DIR"/app/build/"$BUILD_DIR" ]; then
     rm -rf "$DOCKER_ZMK_DIR/app/build/$BUILD_DIR"
+    printf "♻ Removed old build directory befort starting the build process.\n"
+fi
 
-    printf "\n🚧 Starting the build for \"$board MCU ${shield:+($shield keyboard)}\"\n"
-    printf "╰┈┈➤"
-    compile_board "$board" "$shield"
-    printf "\n\n"
-done
-echo
-# grep -v -e "^#" -e "^$" "$DOCKER_ZMK_DIR/app/build/$BUILD_DIR/zephyr/.config" | sort
+compile_board "$board" "$shield"
+
